+++ app/example.cpp
@@ -6,8 +6,6 @@
 
 using namespace Project;
 using namespace Project::delameta::http;
-using delameta::TCP;
-using delameta::Stream;
 using delameta::URL;
 using delameta::info;
 using etl::Ref;
@@ -15,21 +13,21 @@ using etl::Ok;
 using etl::Err;
 
 // define some json rules for some http classes
-JSON_DEFINE(Router, 
-    JSON_ITEM("methods", methods), 
+JSON_DEFINE(Router,
+    JSON_ITEM("methods", methods),
     JSON_ITEM("path", path)
 )
 
-JSON_DEFINE(Error, 
+JSON_DEFINE(Error,
     JSON_ITEM("err", what)
 )
 
-JSON_DEFINE(URL, 
-    JSON_ITEM("url", url), 
-    JSON_ITEM("protocol", protocol), 
-    JSON_ITEM("host", host), 
-    JSON_ITEM("path", path), 
-    JSON_ITEM("full_path", full_path), 
+JSON_DEFINE(URL,
+    JSON_ITEM("url", url),
+    JSON_ITEM("protocol", protocol),
+    JSON_ITEM("host", host),
+    JSON_ITEM("path", path),
+    JSON_ITEM("full_path", full_path),
     JSON_ITEM("queries", queries),
     JSON_ITEM("fragment", fragment)
 )
@@ -40,8 +38,8 @@ struct Foo {
     std::string text;
 };
 
-JSON_DEFINE(Foo, 
-    JSON_ITEM("num", num), 
+JSON_DEFINE(Foo,
+    JSON_ITEM("num", num),
     JSON_ITEM("text", text)
 )
 
@@ -69,7 +67,7 @@ static auto get_token(const RequestReader& req, ResponseWriter&) -> Result<std::
     auto it = req.headers.find("Authentication");
     if (it == req.headers.end()) {
         it = req.headers.find("authentication");
-    } 
+    }
     if (it != req.headers.end()) {
         token = it->second;
     } else {
@@ -99,104 +97,104 @@ HTTP_SETUP(example, app) {
     };
 
     // example: print hello
-    app.Get("/hello", {}, 
+    app.Get("/hello")|
     []() {
         return "Hello world from delameta/" DELAMETA_VERSION;
-    });
+    };
 
     // example: print hello with msg
-    app.Post("/hello", std::tuple{arg::json_item("msg")}, 
+    app.Post("/hello").args(arg::json_item("msg"))|
     [](std::string msg) {
         return "Hello world " + msg;
-    });
+    };
 
-    // example: 
+    // example:
     // - get request param (in this case the body as string_view)
     // - possible error return value
-    app.Post("/body", std::tuple{arg::body},
+    app.Post("/body").args(arg::body)|
     [](std::string_view body) -> Result<std::string_view> {
         if (body.empty()) {
             return etl::Err(Error{StatusBadRequest, "Body is empty"});
         } else {
             return etl::Ok(body);
         }
-    });
+    };
 
-    // example: 
-    // - dependency injection (in this case is authentication token), 
+    // example:
+    // - dependency injection (in this case is authentication token),
     // - arg with default value, it will try to find "add" key in the request headers and request queries
     //   If not found, use the default value
     // - since json rule for Foo is defined and the arg type is json, the request body will be deserialized into foo
     // - new foo will be created and serialized as response body
-    app.Post("/foo", std::tuple{arg::depends(get_token), arg::default_val("add", 20), arg::json},
+    app.Post("/foo").args(arg::depends(get_token), arg::default_val("add", 20), arg::json)|
     [](std::string_view token, int add, Foo foo) -> Foo {
-        return {foo.num + add, foo.text + ": " + std::string(token)}; 
-    });
+        return {foo.num + add, foo.text + ": " + std::string(token)};
+    };
 
-    // example: 
+    // example:
     // - it will find "bar" key in the request headers and request queries
     // - if found, it will be deserialized using custom Http::convert_string_into
-    // - since Http::process_result for Bar is defined, the return value of this function will be used to process the result 
-    app.Get("/bar", std::tuple{arg::arg("bar")}, 
-    [](Bar bar) -> Bar { 
-        return bar; 
-    });
+    // - since Http::process_result for Bar is defined, the return value of this function will be used to process the result
+    app.Get("/bar").args(arg::arg("bar"))|
+    [](Bar bar) -> Bar {
+        return bar;
+    };
 
     // example:
     // multiple methods handler
-    app.route("/methods", {"GET", "POST"}, std::tuple{arg::method},
+    app.route("/methods", {"GET", "POST"}).args(arg::method)|
     [](std::string_view method) {
         if (method == "GET") {
             return "Example GET method";
         } else {
             return "Example POST method";
         }
-    });
+    };
 
     // example: print all routes of this app as json list
-    app.Get("/routes", {},
+    app.Get("/routes")|
     [&]() -> Ref<const std::list<Router>> {
         return etl::ref_const(app.routers);
-    });
+    };
 
     // example: print all headers
-    app.Get("/headers", std::tuple{arg::headers},
+    app.Get("/headers").args(arg::headers)|
     [](decltype(RequestReader::headers) headers) {
         return headers;
-    });
+    };
 
     // example: print all queries
-    app.Get("/queries", std::tuple{arg::queries},
+    app.Get("/queries").args(arg::queries)|
     [](decltype(URL::queries) queries) {
         return queries;
-    });
+    };
 
     // example: print url
-    app.Get("/url", std::tuple{arg::url},
+    app.Get("/url").args(arg::url)|
     [](URL url) {
         return url;
-    });
+    };
 
     // example: print url
-    app.Get("/resolve_url", std::tuple{arg::arg("url")},
+    app.Get("/resolve_url").args(arg::arg("url"))|
     [](std::string url) {
         return URL(url);
-    });
+    };
 
-    // example: redirect to the given path
-    app.route("/redirect", {"GET", "POST", "PUT", "PATCH", "HEAD", "TRACE", "DELETE", "OPTIONS"}, 
-        std::tuple{arg::request, arg::arg("url")}, 
-    [](Ref<const RequestReader> req, std::string url) -> Result<ResponseWriter> {
+    // example: redirect to the given url
+    app.route("/redirect", {"GET", "POST", "PUT", "PATCH", "HEAD", "TRACE", "DELETE", "OPTIONS"})
+    .args(arg::request, arg::arg("url"))
+    .service([](Ref<const RequestReader> req, std::string url) -> Result<ResponseWriter> {
         RequestWriter req_w = *req;
         req_w.url = url;
 
         auto res = TRY(request(std::move(req_w)));
         ResponseWriter res_w = std::move(res);
-    
+
         return Ok(std::move(res_w));
     });
 
-    app.Delete("/delete_route", std::tuple{arg::arg("path")},
+    app.Delete("/delete_route").args(arg::arg("path"))|
     [&](std::string path) -> Result<void> {
         auto it = std::find_if(app.routers.begin(), app.routers.end(), [&path](Router& router) {
             return router.path == path;
@@ -208,5 +206,5 @@ HTTP_SETUP(example, app) {
 
         app.routers.erase(it);
         return Ok();
-    });
+    };
 }