+++ core/stm32_hal/http.cpp
@@ -1,0 +1,9 @@
+#include "delameta/http/http.h"
+
+using namespace Project;
+using namespace delameta;
+using etl::Err;
+
+auto http::Http::Static(std::string path, const std::string& static_dir) -> delameta::Result<void> {
+    return Err("Not implemented");
+}